Update initial overlay
